Fix 049 teleporting to player room
This code would teleport 049 to a room adjacent to a room adjacent to the player room. Well, the player room itself is a room adjacent to a room adjacent to the player room.

There are multiple ways I could've implemented this extra PlayerRoom check but I decided to match the way the code does it further down at around line 1940.