Fix 049 appearing on blue NVG interface while idle
When 049 is idle, the game positions his collider to 0,-500,0, but because ResetEntity isn't called, his collider collides with the floor and remains above ground while his model successfully goes below ground because it has no collision. This results in 049's collider showing up on the blue NVG interface even though he is idle and his model can't be seen.

I also deleted the PositionEntity call on the model because it's redundant: the bottom of 049's case (around line 2145) contains a line that positions the model to wherever the collider is.